+++ lib/factory_sloth/create_call.rb
@@ -1,1 +1,1 @@
-FactorySloth::CreateCall = Struct.new(:name, :line, :column, keyword_init: true)
+FactorySloth::CreateCall = Struct.new(:column, :line, :name, keyword_init: true)